Move window ops to window package
